stable_diffusion_research/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stable-diffusion-research"
version = "0.1.0"
description = "Research implementation of Stable Diffusion for text-to-image generation"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Research Team"}
]
keywords = ["diffusion", "stable-diffusion", "text-to-image", "generative-ai", "deep-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.36.0",
    "diffusers>=0.25.0",
    "accelerate>=0.25.0",
    "datasets>=2.15.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "mlflow>=2.9.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "einops>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
sd-train = "scripts.train:main"
sd-evaluate = "scripts.evaluate:main"
sd-generate = "scripts.generate:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100
